was made the last view for 4 days  forecast
TODO: resolve constraints warnings

+++ weatherAppp/Model/CurrentWeather/Coordinates .swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct Coordinates: Codable {
-    let lon: Float
-    let lat: Float
+    let lon: Double
+    let lat: Double
 }
